Move runtime settings back that are agent specific
