+++ pkg/config/settings/runtime_setting.go
@@ -73,11 +73,11 @@ func GetRuntimeSetting(setting string) (interface{}, error) {
 	return value, nil
 }
 
-// getBool returns the bool value contained in value.
+// GetBool returns the bool value contained in value.
 // If value is a bool, returns its value
 // If value is a string, it converts "true" to true and "false" to false.
 // Else, returns an error.
-func getBool(v interface{}) (bool, error) {
+func GetBool(v interface{}) (bool, error) {
 	// to be cautious, take care of both calls with a string (cli) or a bool (programmaticaly)
 	str, ok := v.(string)
 	if ok {
@@ -88,13 +88,13 @@ func getBool(v interface{}) (bool, error) {
 		case "false":
 			return false, nil
 		default:
-			return false, fmt.Errorf("getBool: bad parameter value provided: %v", str)
+			return false, fmt.Errorf("GetBool: bad parameter value provided: %v", str)
 		}
 
 	}
 	b, ok := v.(bool)
 	if !ok {
-		return false, fmt.Errorf("getBool: bad parameter value provided")
+		return false, fmt.Errorf("GetBool: bad parameter value provided")
 	}
 	return b, nil
 }
